requirements.txt
GitPython==3.1.27
python-dotenv==0.20.0
requests==2.27.1
pipreqs==0.4.11